Updated Drivers, fixed an error in the Shooter Motor Subsystem

+++ src/main/java/frc/robot/subsystems/shooterMotorSubsystem.java
@@ -107,7 +107,7 @@ public void runShooterEncoder() {
 
 }
 
-public int getVelocityError() {
+public double getVelocityError() {
   
   
   System.out.print(" Velocity1 ");
